Add person search typeahead to dashboard #531
