+++ src/app/dashboard/dashboard-routing.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
+import { dossierRoute } from '../shared/components/student-dossier/dossier-route';
 
 const routes: Routes = [
   {
     path: '',
     component: DashboardComponent,
   },
+  dossierRoute,
 ];
 
 @NgModule({
