+++ src/app/home.component.spec.ts
@@ -36,7 +36,6 @@ describe('HomeComponent', () => {
       '/edit-absences',
       '/evaluate-absences',
       '/events',
-      '/person-search',
       '/my-absences',
       '/my-profile',
       '/my-grades',
