+++ src/app/home.component.ts
@@ -24,7 +24,6 @@ export class HomeComponent {
     'edit-absences',
     'evaluate-absences',
     'events',
-    'person-search',
     'my-absences',
     'my-profile',
     'my-grades',
