--- src/app/person-search/components/person-search-header/person-search-header.component.html
@@ -1,10 +1,0 @@
-<div class="px-3 py-3 border-bottom">
-  <label class="form-label">{{ 'person-search.name' | translate }}</label>
-  <erz-typeahead
-    [typeaheadService]="studentsRestService"
-    [value]="filter.student"
-    (valueChange)="$event && navigateToDossier($event)"
-  >
-  </erz-typeahead>
-</div>
-<p class="mt-3 ps-3">{{ 'person-search.description' | translate }}</p>
