--- src/app/person-search/components/person-search/person-search.component.html
@@ -1,1 +1,0 @@
-<router-outlet></router-outlet>
