--- src/app/person-search/components/person-search/person-search.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'erz-search-person',
-  templateUrl: './person-search.component.html',
-  styleUrls: ['./person-search.component.scss'],
-})
-export class PersonSearchComponent {
-  constructor() {}
-}
