+++ src/assets/locales/de-CH.json
@@ -334,11 +334,6 @@
       }
     }
   },
-  "person-search": {
-    "title": "Personensuche",
-    "description": "Hier können Sie nach Ihren Lernenden suchen, um mit Klick auf den Namen direkt zum Dossier zu gelangen.",
-    "name": "Name"
-  },
   "my-profile": {
     "title": "Mein Profil",
     "show": {
